refactor: useWeekDates 훅과 무관한 함수 분리 #32
